Fixed Checking and Default
Added in checking the excel input file to stop nan's as well as  updated the default RGB values. Lastly, I made a change from just changing alpha's to 0 that were not touched by the deer to making them be black pixels, thus increasing the contrast of the more transparent pixels.